move Crawler to inside Mwcrawler scope
